device responsivity - fix, task - Reset, Verify, ResetPassword, tailwind.config.js
I made modifications with my pages to incorporate responsivity. This pull request contains modifications to the tailwind.config.js along with the new Reset, Verify, and ResetPassword pages. Reset.jsx contains information on the first page on the Figma; it requests for the user's email and is meant to send a verification code to the user. Verify.jsx gives the user a timer to authenticate the code. ResetPassword asks the user to confirm their password changes. I modified the tailwind.config.js files to add different screen sizes to add device responsitivity.